+++ packages/cli/src/lib/commands/init/constants.ts
@@ -1,4 +1,4 @@
-import { FlowExamples,GhWorkflowExamples } from './types';
+import { FlowExamples, GhWorkflowExamples, BudgetsExamples } from './types';
 
 export const SETUP_CONFIRM_MESSAGE = 'user-flow CLI is set up now! 🎉';
 export const FlowExampleMap: Record<FlowExamples, string> = {
@@ -7,3 +7,19 @@ export const FlowExampleMap: Record<FlowExamples, string> = {
 export const GhWorkflowExampleMap: Record<GhWorkflowExamples, string> = {
   'basic-workflow': 'user-flow-ci.yml'
 };
+export const BudgetsExampleMap: Record<BudgetsExamples, string> = {
+  'budgets': 'lh-budgets.json'
+};
+export const TransferSizeResourceTypeBudgets: string[] = [
+  'xhr',
+  'script',
+  'stylesheet',
+  'font',
+  'document',
+  'third-party'
+];
+export const RequestCountResourceTypeBudgets: string[] = [
+  'xhr',
+  'script',
+  'stylesheet'
+];